Repo clean ;)
Utilis split, hardcoded model fix,
