+++ model.py
@@ -1,6 +1,7 @@
 import torch.nn as nn
 import torch
 import numpy as np
+import random 
 
 
 class Res_block(nn.Module):
